Extract Code Frame Component
