+++ packages/react-dev-overlay/src/internal/components/CodeFrame/index.tsx
@@ -1,0 +1,1 @@
+export { CodeFrame } from './CodeFrame'
